Refactor: 리뷰 스타일링 코드 reviewStyle로 이동
